requirements.txt
visdom
dominate
matplotlib
numpy
Pillow
scipy
tqdm